+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # **Changelog**
-### Version [1.13.1] - 2026-06-21
+### Version [1.13.3] - 2026-06-21
 #### Updated
 * Updated the supported interface version to 120007.